Replace vxcan.dev and vxcan.id with vxcan.device
It Looks like the original intent of vxcan.id was that it could be used to
specify the numerical identifier of an interface both in and out of
containers, so that you could connect `can2` on your host to `vcan2`
in the container (different prefix, same number). However, as noted in
the README, this isn't actually possible due to how docker enumerates
the network devices in containers.

Given that situation, there's really no reason to specify the host device
piecewise like this, it's simpler to work with if you just specify the
full name of the can device on the host, and the prefix for the container.